refactor(core): move the run vocabulary out of rig-run into rig-core
transcript helpers, OutputMode + output-policy helpers, policy data,
prepare_request/PreparedRequest, RunSpec, PromptResponse/CompletionCall/
PromptError and the streamed-turn assembler now live in rig-core
(rig_core::{transcript, completion::{output, policy, prepare, spec,
response}, streaming::assemble}). rig-run keeps AgentRun and re-exports the
moved modules at its old paths so rig-agent builds unchanged mid-series.

AgentRun::from_spec and AgentRun::with_validated_history (which straddled
the split) move into run.rs; DEFAULT_OUTPUT_RETRIES becomes
RunSpec::DEFAULT_OUTPUT_RETRIES; the streamed tests that drive an AgentRun
move into run.rs (14 tests) while the assembler-only tests stay in rig-core.
