+++ crates/rig-core/src/completion/output.rs
