+++ crates/rig-core/src/lib.rs
@@ -173,6 +173,7 @@ pub mod streaming;
 #[cfg_attr(docsrs, doc(cfg(feature = "test-utils")))]
 pub mod test_utils;
 pub mod tool;
+pub mod transcript;
 pub mod transcription;
 pub mod vector_store;
 pub mod wasm_compat;
