+++ crates/rig-run/src/run.rs
@@ -64,7 +64,7 @@
 //! # }
 //! ```
 
-pub use crate::output_mode::OutputMode;
+pub use rig_core::completion::output::OutputMode;
 
 use std::collections::{BTreeMap, BTreeSet};
 
@@ -92,20 +92,20 @@ use rig_core::message::{
     AssistantContent, ToolCall, ToolChoice, ToolResult, ToolResultContent, UserContent,
 };
 
-use crate::{
-    error::PromptError,
-    policy::{InvalidToolCallAction, InvalidToolCallContext, RetryRequest},
-    response::{CompletionCall, PromptResponse},
-    transcript::{
-        TOOL_NOT_EXECUTED_DUE_TO_INVALID_PEER, assistant_text_from_choice, build_full_history,
-        build_history_for_request, invalid_tool_retry_user_message, is_empty_assistant_turn,
-        tool_result_message, turn_delivered_no_answer,
-    },
-};
 use rig_core::completion::{Message, ResponseIdentity, Usage};
+use rig_core::completion::{
+    policy::{InvalidToolCallAction, InvalidToolCallContext, RetryRequest},
+    response::{CompletionCall, PromptError, PromptResponse},
+    spec::RunSpec,
+};
 use rig_core::json_utils;
+use rig_core::transcript::{
+    TOOL_NOT_EXECUTED_DUE_TO_INVALID_PEER, TranscriptError, assistant_text_from_choice,
+    build_full_history, build_history_for_request, invalid_tool_retry_user_message,
+    is_empty_assistant_turn, tool_result_message, turn_delivered_no_answer, validate_canonical,
+};
 
-pub use crate::streamed::{
+pub use rig_core::streaming::assemble::{
     PartialStreamedTurn, StreamedInvalidToolCall, StreamedResolution, StreamedTurn,
     StreamedTurnAssembler, StreamedTurnEvent,
 };
@@ -161,11 +161,6 @@ enum ValidatedInvalidToolCallAction {
     Repair { tool_name: String },
     Skip { reason: String },
 }
-
-/// Default number of times Tool output mode re-prompts the model for valid
-/// structured output before finalizing best-effort (see #1928). Mirrors
-/// pydantic-ai's default output-retry budget of 1.
-pub const DEFAULT_OUTPUT_RETRIES: usize = 1;
 
 /// What a driver must do next to advance an [`AgentRun`].
 ///
@@ -246,7 +241,7 @@ pub struct ModelTurn {
 
 impl ModelTurn {
     /// The one blessed conversion from a provider response to a protocol
-    /// turn, given the [`PreparedRequest`](crate::prepare::PreparedRequest)
+    /// turn, given the [`PreparedRequest`](rig_core::completion::prepare::PreparedRequest)
     /// the call was prepared from.
     ///
     /// Every driver — rig-agent's futures runner and any external systems
@@ -260,7 +255,7 @@ impl ModelTurn {
     /// a raw provider field.
     pub fn from_response(
         resp: &CompletionResponse,
-        prepared: &crate::prepare::PreparedRequest,
+        prepared: &rig_core::completion::prepare::PreparedRequest,
     ) -> Self {
         Self::from_response_parts(
             resp,
@@ -1609,7 +1604,7 @@ impl AgentRun {
     }
 
     // ── Streamed-turn entry points ──────────────────────────────────────
-    // Paired with [`crate::streamed::StreamedTurnAssembler`]; see that module's
+    // Paired with [`rig_core::streaming::assemble::StreamedTurnAssembler`]; see that module's
     // docs for the full driving protocol.
 
     /// Record one provider completion call for a streamed turn.
@@ -1651,7 +1646,7 @@ impl AgentRun {
     }
 
     /// The recovery-hook context for an invalid tool call surfaced
-    /// mid-stream by a [`crate::streamed::StreamedTurnAssembler`].
+    /// mid-stream by a [`rig_core::streaming::assemble::StreamedTurnAssembler`].
     pub fn streamed_invalid_tool_call_context(
         &self,
         partial: &PartialStreamedTurn,
@@ -1870,6 +1865,46 @@ impl AgentRun {
             self.full_history(),
             format!("agent run driver protocol violation: {reason}"),
         )
+    }
+}
+
+impl AgentRun {
+    /// Build a run from a [`RunSpec`], a prompt and an optional prior history.
+    ///
+    /// Applies the spec's budget, invalid-call retries, output validation and
+    /// tool choice; everything else in the spec is request-shaping the driver
+    /// reads when it prepares each model call.
+    pub fn from_spec(
+        spec: &RunSpec,
+        prompt: impl Into<Message>,
+        history: Option<Vec<Message>>,
+    ) -> Self {
+        let mut run = AgentRun::new(prompt)
+            .max_turns(spec.effective_max_turns())
+            .max_invalid_tool_call_retries(spec.max_invalid_tool_call_retries)
+            .with_output_validation(spec.output_schema.clone(), RunSpec::DEFAULT_OUTPUT_RETRIES);
+        if let Some(history) = history {
+            run = run.with_history(history);
+        }
+        if let Some(tool_choice) = spec.tool_choice.clone() {
+            run = run.with_tool_choice(tool_choice);
+        }
+        if let Some(name) = spec.output_tool_name.clone() {
+            run = run.with_output_tool_name(name);
+        }
+        run
+    }
+}
+
+impl AgentRun {
+    /// [`with_history`](AgentRun::with_history), rejecting a history that is
+    /// not a canonical transcript. Use this when the history comes from
+    /// outside the protocol (a memory backend, a resumed run from another
+    /// process); `with_history` stays unchecked for callers that built the
+    /// history themselves.
+    pub fn with_validated_history(self, history: Vec<Message>) -> Result<Self, TranscriptError> {
+        validate_canonical(&history)?;
+        Ok(self.with_history(history))
     }
 }
 
@@ -3534,5 +3569,560 @@ mod tests {
         let legacy: CompletionCall =
             serde_json::from_value(unset).expect("a call without a raw key still loads");
         assert_eq!(legacy.raw, serde_json::Value::Null);
+    }
+
+    #[test]
+    fn from_spec_matches_the_builder_chain() {
+        let spec = RunSpec {
+            max_turns: Some(4),
+            max_invalid_tool_call_retries: 2,
+            tool_choice: Some(ToolChoice::Auto),
+            ..RunSpec::new()
+        };
+        let via_spec = AgentRun::from_spec(&spec, "hi", None);
+        let via_chain = AgentRun::new("hi")
+            .max_turns(4)
+            .max_invalid_tool_call_retries(2)
+            .with_output_validation(None, RunSpec::DEFAULT_OUTPUT_RETRIES)
+            .with_tool_choice(ToolChoice::Auto);
+        assert_eq!(
+            serde_json::to_value(&via_spec).expect("serialize"),
+            serde_json::to_value(&via_chain).expect("serialize")
+        );
+    }
+
+    fn assistant(content: Vec<AssistantContent>) -> Message {
+        Message::Assistant { id: None, content }
+    }
+
+    #[test]
+    fn with_validated_history_gates_construction() {
+        let bad = vec![
+            assistant(vec![AssistantContent::text("a")]),
+            assistant(vec![AssistantContent::text("b")]),
+        ];
+        assert!(AgentRun::new("x").with_validated_history(bad).is_err());
+        assert!(
+            AgentRun::new("x")
+                .with_validated_history(vec![Message::user("ok")])
+                .is_ok()
+        );
+    }
+}
+
+#[cfg(test)]
+mod streamed_tests {
+    use super::*;
+    use std::collections::BTreeSet;
+
+    use rig_core::id::InternalCallId;
+    use rig_core::message::{Text, ToolFunction, ToolResultContent, UserContent};
+    use rig_core::streaming::{StreamedAssistantContent, ToolCallDeltaContent};
+    use rig_core::test_utils::mock_final;
+    use serde_json::json;
+
+    fn tool_names(names: &[&str]) -> BTreeSet<String> {
+        names.iter().map(|name| (*name).to_string()).collect()
+    }
+
+    fn assembler() -> StreamedTurnAssembler {
+        StreamedTurnAssembler::new(tool_names(&["add"]), tool_names(&["add"]))
+    }
+
+    fn text_item(text: &str) -> StreamedAssistantContent {
+        StreamedAssistantContent::Text(Text::new(text.to_string()))
+    }
+
+    /// Deterministic test-only internal id derived from a wire id, so a
+    /// delta and its completed call correlate without threading state.
+    fn iid_for(id: &str) -> InternalCallId {
+        let mut hash: u64 = 0xcbf2_9ce4_8422_2325;
+        for byte in id.bytes() {
+            hash ^= u64::from(byte);
+            hash = hash.wrapping_mul(0x100_0000_01b3);
+        }
+        InternalCallId::from_raw(hash | 1).expect("non-zero")
+    }
+
+    fn tool_call(id: &str, name: &str) -> ToolCall {
+        // The provider-boundary shape: the wire id becomes both the durable
+        // id and the provider correlator.
+        ToolCall::from_wire(id, ToolFunction::new(name.to_string(), json!({"x": 1})))
+    }
+
+    fn tool_call_item(id: &str, name: &str) -> StreamedAssistantContent {
+        StreamedAssistantContent::ToolCall {
+            tool_call: tool_call(id, name),
+            internal_call_id: iid_for(id),
+        }
+    }
+
+    fn name_delta(id: &str, name: &str) -> StreamedAssistantContent {
+        StreamedAssistantContent::ToolCallDelta {
+            internal_call_id: iid_for(id),
+            content: ToolCallDeltaContent::Name(name.to_string()),
+        }
+    }
+
+    fn args_delta(id: &str, arguments: &str) -> StreamedAssistantContent {
+        StreamedAssistantContent::ToolCallDelta {
+            internal_call_id: iid_for(id),
+            content: ToolCallDeltaContent::Delta(arguments.to_string()),
+        }
+    }
+
+    fn expect_invalid(events: Vec<StreamedTurnEvent>) -> StreamedInvalidToolCall {
+        match events.into_iter().next() {
+            Some(StreamedTurnEvent::InvalidToolCall(invalid)) => *invalid,
+            other => panic!("expected InvalidToolCall, got {other:?}"),
+        }
+    }
+
+    #[test]
+    fn streamed_run_completes_a_tool_roundtrip() {
+        let mut run = AgentRun::new("add things").max_turns(2);
+
+        // Turn 1: the model streams one tool call.
+        let AgentRunStep::CallModel { .. } = run.next_step().expect("next_step") else {
+            panic!("expected CallModel");
+        };
+        let mut asm = assembler();
+        assert!(
+            asm.ingest(&tool_call_item("tc_1", "add"))
+                .expect("ingest should succeed")
+                .is_empty()
+        );
+        let usage = Usage {
+            input_tokens: 5,
+            output_tokens: 7,
+            total_tokens: 12,
+            ..Usage::new()
+        };
+        run.record_streamed_completion_call(
+            usage,
+            rig_core::completion::ResponseIdentity::default(),
+            None,
+            serde_json::Value::Null,
+        )
+        .expect("record should succeed");
+        let final_choice = vec![AssistantContent::ToolCall(tool_call("tc_1", "add"))];
+        run.streamed_turn(asm.finish(Some("msg_1".to_string()), &final_choice))
+            .expect("streamed_turn should succeed");
+
+        let AgentRunStep::CallTools { calls } = run.next_step().expect("next_step") else {
+            panic!("expected CallTools");
+        };
+        assert_eq!(calls.len(), 1);
+        assert_eq!(calls[0].internal_call_id, Some(iid_for("tc_1")));
+        run.tool_results(vec![UserContent::tool_result(
+            "tc_1",
+            "add",
+            vec![ToolResultContent::text("2")],
+        )])
+        .expect("tool_results should succeed");
+
+        // Turn 2: plain text finishes the run.
+        let AgentRunStep::CallModel { .. } = run.next_step().expect("next_step") else {
+            panic!("expected CallModel");
+        };
+        let asm = assembler();
+        run.record_streamed_completion_call(
+            Usage::new(),
+            rig_core::completion::ResponseIdentity::default(),
+            None,
+            serde_json::Value::Null,
+        )
+        .expect("record should succeed");
+        let final_choice = vec![AssistantContent::text("done")];
+        run.streamed_turn(asm.finish(None, &final_choice))
+            .expect("streamed_turn should succeed");
+
+        let AgentRunStep::Done(response) = run.next_step().expect("next_step") else {
+            panic!("expected Done");
+        };
+        assert_eq!(response.output, "done");
+        assert_eq!(response.usage, usage);
+        assert_eq!(response.completion_calls.len(), 2);
+        assert_eq!(response.completion_calls[0].usage, usage);
+        assert_eq!(response.completion_calls[1].usage, Usage::new());
+        // prompt, assistant tool call, tool result, final assistant text
+        assert_eq!(
+            response
+                .messages
+                .expect("messages should be recorded")
+                .len(),
+            4
+        );
+    }
+
+    #[test]
+    fn streamed_invalid_tool_call_retry_rolls_back_with_partial_turn() {
+        let mut run = AgentRun::new("use the tool")
+            .max_turns(2)
+            .max_invalid_tool_call_retries(1);
+        run.next_step().expect("next_step");
+
+        let mut asm = assembler();
+        asm.ingest(&text_item("thinking ")).expect("ingest");
+        let invalid = expect_invalid(
+            asm.ingest(&tool_call_item("tc_1", "default_api"))
+                .expect("ingest should succeed"),
+        );
+        let partial = asm.partial_turn(Some("msg_1".to_string()));
+        assert_eq!(partial.text.as_deref(), Some("thinking "));
+
+        let context = run.streamed_invalid_tool_call_context(&partial, &invalid);
+        assert!(context.is_streaming);
+        assert_eq!(context.tool_name, "default_api");
+        assert_eq!(context.internal_call_id, Some(iid_for("tc_1")));
+
+        let resolution = run
+            .resolve_streamed_invalid_tool_call(
+                &partial,
+                &invalid,
+                InvalidToolCallAction::retry("use add instead"),
+            )
+            .expect("retry should be accepted");
+        assert!(matches!(
+            resolution,
+            StreamedResolution::TurnAbandoned {
+                skipped_tool_result: None
+            }
+        ));
+        asm.resolve_pending_invalid(&resolution);
+
+        // Usage from the drained stream is recorded after the rollback.
+        run.record_streamed_completion_call(
+            Usage::new(),
+            rig_core::completion::ResponseIdentity::default(),
+            None,
+            serde_json::Value::Null,
+        )
+        .expect("record after rollback should succeed");
+
+        // The rollback appended the partial assistant turn and feedback.
+        assert_eq!(run.messages().len(), 3);
+        let AgentRunStep::CallModel { turn, .. } = run.next_step().expect("next_step") else {
+            panic!("expected CallModel retry");
+        };
+        assert_eq!(turn, 2);
+    }
+
+    #[test]
+    fn streamed_invalid_tool_call_stop_leaves_run_terminal() {
+        let mut run = AgentRun::new("use the tool");
+        run.next_step().expect("next_step");
+
+        let mut asm = assembler();
+        let invalid = expect_invalid(
+            asm.ingest(&tool_call_item("tc_1", "default_api"))
+                .expect("ingest should succeed"),
+        );
+        let partial = asm.partial_turn(Some("msg_1".to_string()));
+
+        let err = run
+            .resolve_streamed_invalid_tool_call(
+                &partial,
+                &invalid,
+                InvalidToolCallAction::stop("operator stop"),
+            )
+            .expect_err("stop should cancel the run");
+        assert!(matches!(
+            err,
+            PromptError::PromptCancelled { reason, .. } if reason == "operator stop"
+        ));
+
+        let err = run
+            .next_step()
+            .expect_err("a stopped streamed run must remain terminal");
+        assert!(matches!(
+            err,
+            PromptError::PromptCancelled { reason, .. }
+                if reason.contains("next_step called after the run already failed")
+        ));
+    }
+
+    #[test]
+    fn streamed_invalid_tool_call_retry_cannot_emit_call_past_total_budget() {
+        let mut run = AgentRun::new("use the tool")
+            .max_turns(1)
+            .max_invalid_tool_call_retries(1);
+        run.next_step().expect("initial model call");
+
+        let mut asm = assembler();
+        let invalid = expect_invalid(
+            asm.ingest(&tool_call_item("tc_1", "default_api"))
+                .expect("ingest should succeed"),
+        );
+        let partial = asm.partial_turn(Some("msg_1".to_string()));
+        let resolution = run
+            .resolve_streamed_invalid_tool_call(
+                &partial,
+                &invalid,
+                InvalidToolCallAction::retry("use add instead"),
+            )
+            .expect("retry resolution should be accepted");
+        assert!(matches!(
+            resolution,
+            StreamedResolution::TurnAbandoned {
+                skipped_tool_result: None
+            }
+        ));
+        run.record_streamed_completion_call(
+            Usage::new(),
+            rig_core::completion::ResponseIdentity::default(),
+            None,
+            serde_json::Value::Null,
+        )
+        .expect("completion call should be recorded");
+        assert_eq!(run.completion_calls().len(), 1);
+
+        let err = run
+            .next_step()
+            .expect_err("retry must not emit a second model call");
+        assert!(matches!(
+            err,
+            PromptError::MaxTurnsError { max_turns: 1, .. }
+        ));
+        assert_eq!(run.turn(), 1);
+    }
+
+    #[test]
+    fn streamed_invalid_tool_call_skip_returns_synthetic_result() {
+        let mut run = AgentRun::new("use the tool").max_turns(2);
+        run.next_step().expect("next_step");
+
+        let mut asm = assembler();
+        let invalid = expect_invalid(
+            asm.ingest(&tool_call_item("tc_1", "default_api"))
+                .expect("ingest should succeed"),
+        );
+        let partial = asm.partial_turn(None);
+
+        let resolution = run
+            .resolve_streamed_invalid_tool_call(
+                &partial,
+                &invalid,
+                InvalidToolCallAction::skip("not available"),
+            )
+            .expect("skip should be accepted");
+        let StreamedResolution::TurnAbandoned {
+            skipped_tool_result: Some(tool_result),
+        } = &resolution
+        else {
+            panic!("expected skipped tool result");
+        };
+        assert_eq!(tool_result.call, "tc_1");
+    }
+
+    #[test]
+    fn streamed_invalid_name_delta_repair_replays_buffered_arguments() {
+        let mut run = AgentRun::new("use the tool").max_turns(2);
+        run.next_step().expect("next_step");
+
+        let mut asm = assembler();
+        asm.ingest(&args_delta("tc_1", "{\"x\":1}"))
+            .expect("ingest should succeed");
+        let invalid = expect_invalid(
+            asm.ingest(&name_delta("tc_1", "default_api"))
+                .expect("ingest should succeed"),
+        );
+        assert_eq!(invalid.args.as_deref(), Some("{\"x\":1}"));
+
+        let partial = asm.partial_turn(None);
+        let resolution = run
+            .resolve_streamed_invalid_tool_call(
+                &partial,
+                &invalid,
+                InvalidToolCallAction::repair("add"),
+            )
+            .expect("repair should be accepted");
+        assert!(matches!(
+            resolution,
+            StreamedResolution::Repaired { ref tool_name } if tool_name == "add"
+        ));
+
+        let events = asm.resolve_pending_invalid(&resolution);
+        let contents: Vec<_> = events
+            .iter()
+            .map(|event| match event {
+                StreamedTurnEvent::EmitToolCallDelta { content, .. } => content.clone(),
+                other => panic!("expected EmitToolCallDelta, got {other:?}"),
+            })
+            .collect();
+        assert_eq!(
+            contents,
+            vec![
+                ToolCallDeltaContent::Name("add".to_string()),
+                ToolCallDeltaContent::Delta("{\"x\":1}".to_string()),
+            ]
+        );
+    }
+
+    #[test]
+    fn streamed_turn_rejects_unknown_tool_calls_fail_fast() {
+        let mut run = AgentRun::new("use the tool");
+        run.next_step().expect("next_step");
+
+        let turn = StreamedTurn {
+            message_id: None,
+            choice: vec![AssistantContent::ToolCall(tool_call("tc_1", "unknown"))],
+            executable_tool_names: tool_names(&["add"]),
+            allowed_tool_names: tool_names(&["add"]),
+            internal_call_ids: Vec::new(),
+            finish_reason: None,
+        };
+        let err = run
+            .streamed_turn(turn)
+            .expect_err("unknown tool should fail fast");
+        assert!(matches!(
+            err,
+            PromptError::UnknownToolCall { tool_name, .. } if tool_name == "unknown"
+        ));
+    }
+
+    #[test]
+    fn streamed_completion_call_record_requires_a_model_call() {
+        // A fresh run has emitted no CallModel: recording must be rejected
+        // even though the machine is in its initial PreparingRequest state.
+        let mut run = AgentRun::new("hello");
+        let err = run
+            .record_streamed_completion_call(
+                Usage::new(),
+                rig_core::completion::ResponseIdentity::default(),
+                None,
+                serde_json::Value::Null,
+            )
+            .expect_err("recording before any model call must be rejected");
+        assert!(matches!(err, PromptError::PromptCancelled { .. }));
+
+        // The run stays drivable.
+        run.next_step().expect("next_step should still succeed");
+        run.record_streamed_completion_call(
+            Usage::new(),
+            rig_core::completion::ResponseIdentity::default(),
+            None,
+            serde_json::Value::Null,
+        )
+        .expect("recording during a pending model call succeeds");
+    }
+
+    #[test]
+    fn duplicate_tool_call_ids_keep_distinct_internal_ids_through_the_run() {
+        let mut run = AgentRun::new("do both").max_turns(2);
+        run.next_step().expect("next_step");
+
+        let mut asm = assembler();
+        asm.ingest(&StreamedAssistantContent::ToolCall {
+            tool_call: tool_call("tc_1", "add"),
+            internal_call_id: iid_for("a"),
+        })
+        .expect("ingest should succeed");
+        asm.ingest(&StreamedAssistantContent::ToolCall {
+            tool_call: tool_call("tc_1", "add"),
+            internal_call_id: iid_for("b"),
+        })
+        .expect("ingest should succeed");
+        run.record_streamed_completion_call(
+            Usage::new(),
+            rig_core::completion::ResponseIdentity::default(),
+            None,
+            serde_json::Value::Null,
+        )
+        .expect("record should succeed");
+
+        let final_choice = vec![
+            AssistantContent::ToolCall(tool_call("tc_1", "add")),
+            AssistantContent::ToolCall(tool_call("tc_1", "add")),
+        ];
+        run.streamed_turn(asm.finish(None, &final_choice))
+            .expect("streamed_turn should succeed");
+
+        // The internal IDs survive in the run state itself: a serde round
+        // trip must keep both calls distinguishable.
+        let serialized = serde_json::to_string(&run).expect("serialize");
+        let mut restored: AgentRun = serde_json::from_str(&serialized).expect("deserialize");
+        let AgentRunStep::CallTools { calls } = restored.next_step().expect("next_step") else {
+            panic!("expected CallTools");
+        };
+        assert_eq!(calls.len(), 2);
+        assert_eq!(calls[0].internal_call_id, Some(iid_for("a")));
+        assert_eq!(calls[1].internal_call_id, Some(iid_for("b")));
+    }
+
+    #[test]
+    fn streamed_turn_records_the_completion_call_when_the_driver_did_not() {
+        let mut run = AgentRun::new("hello");
+        run.next_step().expect("next_step");
+
+        let asm = assembler();
+        let final_choice = vec![AssistantContent::text("done")];
+        run.streamed_turn(asm.finish(None, &final_choice))
+            .expect("streamed_turn should succeed");
+
+        // Exactly one CompletionCall per model call, even without an explicit
+        // record; usage is simply unreported.
+        assert_eq!(run.completion_calls().len(), 1);
+        assert_eq!(run.completion_calls()[0].usage, Usage::new());
+    }
+
+    #[test]
+    fn streamed_completion_call_is_recorded_once_per_turn() {
+        let mut run = AgentRun::new("hello");
+        run.next_step().expect("next_step");
+
+        run.record_streamed_completion_call(
+            Usage::new(),
+            rig_core::completion::ResponseIdentity::default(),
+            None,
+            serde_json::Value::Null,
+        )
+        .expect("first record succeeds");
+        let err = run
+            .record_streamed_completion_call(
+                Usage::new(),
+                rig_core::completion::ResponseIdentity::default(),
+                None,
+                serde_json::Value::Null,
+            )
+            .expect_err("second record for the same turn must be rejected");
+        assert!(matches!(err, PromptError::PromptCancelled { .. }));
+        assert_eq!(run.completion_calls().len(), 1);
+    }
+
+    #[test]
+    fn streamed_run_serde_round_trips_while_tools_pend() {
+        let mut run = AgentRun::new("add things").max_turns(2);
+        run.next_step().expect("next_step");
+
+        let mut asm = assembler();
+        asm.ingest(&tool_call_item("tc_1", "add"))
+            .expect("ingest should succeed");
+        run.record_streamed_completion_call(
+            Usage::new(),
+            rig_core::completion::ResponseIdentity::default(),
+            None,
+            serde_json::Value::Null,
+        )
+        .expect("record should succeed");
+        let final_choice = vec![AssistantContent::ToolCall(tool_call("tc_1", "add"))];
+        run.streamed_turn(asm.finish(None, &final_choice))
+            .expect("streamed_turn should succeed");
+        run.next_step().expect("CallTools step");
+
+        let serialized = serde_json::to_string(&run).expect("serialize mid-run");
+        let mut restored: AgentRun =
+            serde_json::from_str(&serialized).expect("deserialize mid-run");
+        restored
+            .tool_results(vec![UserContent::tool_result(
+                "tc_1",
+                "add",
+                vec![ToolResultContent::text("2")],
+            )])
+            .expect("tool_results should succeed");
+        assert!(matches!(
+            restored.next_step().expect("next turn"),
+            AgentRunStep::CallModel { turn: 2, .. }
+        ));
     }
 }
